% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesFactor.R
\name{BayesFactor}
\alias{BayesFactor}
\title{Computes Bayes Factor of Two Partial M-Probit Models}
\usage{
BayesFactor(m1, m2, log = FALSE)
}
\arguments{
\item{m1}{(required) partial m-probit model 1 estimated with chib95=TRUE}

\item{m2}{(required) partial m-probit model 2 estimated with chib95=TRUE}

\item{log}{report BayesFactor on log-scale?}
}
\value{
A scalar estimate of the Bayes factor.
}
\description{
\code{BayesFactor} computes the Bayes factor given two partial m-probit posteriors for which a marginal likelihood has been computed.
}
\references{
Marbach, Moritz. 2016. `Analyzing Decision Records from Committees.'' Working Paper.

	Chib, Siddhartha. 1995. Marginal Likelihood from the Gibbs Output. Journal of the American Statistical Association 90(432), 1313–1321.
}

