% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consilium-package.R
\docType{package}
\name{consilium-package}
\alias{consilium}
\alias{consilium-package}
\title{Estimating Multivariate Probit Models with Partial Observability}
\description{
Estimates multivariate probit models with partial
observability (partial m-probits) using a double-augmented Gibbs sampler and maximum likelihood. 
When using this package, please cite the paper and the package, see \code{citation("consilium")}.
}
\details{
Feedback is very welcome!


\tabular{ll}{
	Package: \tab consilium\cr
	Type: \tab Package\cr
	Version: \tab 0.0.0.9000\cr
	Date: \tab 2016-11-01\cr
	License: \tab  GPL-3\cr
	}
}
\author{
Moritz Marbach \email{marbach@ipz.uzh.ch}
}
\references{
Marbach, Moritz. 2016. `Analyzing Decision Records from Committees.'' Working Paper.
}

