% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VoteProfiles.R
\name{VoteProfiles}
\alias{VoteProfiles}
\title{Computes All Vote Profiles Given a Veto-Majority Voting Rule}
\usage{
VoteProfiles(R, V = 0, I)
}
\arguments{
\item{R}{(required) the number of members that have to agree 
to pass a proposal excluding the members that have veto power.}

\item{V}{the number of members that have to agree to pass a proposal.}

\item{I}{(required) the number of members voting.}
}
\value{
A \code{list} object with two matrices that list all vote profiles 
	which imply the adoption (\code{yescols}) or rejection (\code{nocols}) of a proposal.  
 A column refers to a voting member and a row to a vote profile. 

Notice that this function is time- and memory intensive if \code{I} is large.
}
\description{
\code{VoteProfiles} computes matrices of vote profiles given a voting rule.
}

